Use JSON parsing for VisionSubsystem (#32)
* Turn yaw into Integer

* Swap vision subsystem over to using the json parsing

* Make yaw a double instead

+++ Competition/src/main/java/competition/subsystems/vision/VisionData.java
@@ -1,19 +1,19 @@
 package competition.subsystems.vision;
 
 public class VisionData {
-    protected int targetYaw;
+    protected Double targetYaw;
 
     /**
      * @return the targetYaw
      */
-    public int getTargetYaw() {
+    public Double getTargetYaw() {
         return targetYaw;
     }
 
     /**
      * @param targetYaw the targetYaw to set
      */
-    public void setTargetYaw(int targetYaw) {
+    public void setTargetYaw(Double targetYaw) {
         this.targetYaw = targetYaw;
     }
 }
